Move Timer class from logging.cpp to new time.h
